Complete lesson 3
Create a new detail activity that shows the forecast string and allows sharing it
through a share menu item.
We also created a Settings activity so the user can select and store location and
temperature units settings.
We added a new Map menu item that will map the current saved location.